fix tests

+++ python/mxnet/image/image.py
@@ -224,9 +224,9 @@ def resize_short(src, size, interp=2):
     """
     h, w, _ = src.shape
     if h > w:
-        new_h, new_w = size * h / w, size
+        new_h, new_w = int(size * h / w), int(size)
     else:
-        new_h, new_w = size, size * w / h
+        new_h, new_w = int(size), int(size * w / h)
     return imresize(src, new_w, new_h, interp=_get_interp_method(interp, (h, w, new_h, new_w)))
